Implement to_from_variant_impls
