+++ crates/macros/cgp-macro-core/src/types/cgp_data/mod.rs
@@ -1,11 +1,15 @@
 mod derive_builder;
 mod derive_extractor;
+mod derive_from_variant;
 mod derive_has_field;
 mod derive_has_fields;
 mod record;
+mod variant;
 
 pub use derive_builder::*;
 pub use derive_extractor::*;
+pub use derive_from_variant::*;
 pub use derive_has_field::*;
 pub use derive_has_fields::*;
 pub use record::*;
+pub use variant::*;
