+++ crates/macros/cgp-macro-lib/src/entrypoints/cgp_data.rs
@@ -8,7 +8,7 @@ pub fn derive_cgp_data(body: TokenStream) -> syn::Result<TokenStream> {
 
     match item {
         Item::Struct(item_struct) => derive_cgp_record_from_struct(item_struct),
-        Item::Enum(item_enum) => derive_cgp_variant_from_enum(&item_enum),
+        Item::Enum(item_enum) => derive_cgp_variant_from_enum(item_enum),
         _ => Err(Error::new_spanned(
             item,
             "expect body to be either a struct or enum",
